requirements.txt
openai==2.12.0
streamlit==1.41.1
python-dotenv==1.0.1
